Closing issue #50, implemented a wrapper around the server ajax call to ensure several constraints. If DEBUG is set in config.py a stack-trace will be returned for any Python version above 2.3.
Additional attributes added to config.py, update according to config_template.py. The server will now always return json for any query as long as ajax.cgi succeeds to be executed (as previously mentioned, anything above 2.3 is fine for this).

Some exits inside of the new ajaxserver.py will be necessary, but for now we are pretty much safe. It is now forbidden to use exit inside ANY function since this will break protocol.