SVG Download
BADLY BUGGED. At this point, there's two options, both bad - and I took the worse one :) as a quicker hack.

a) Use the same file for all users. Whenever anyone displays a document, these svg files are overwritten. Thus, you might very easily get the SVG of a document that has nothing to do with what you have on your screen.

b) Make a random hash, and save the SVG files under that; communicate the hash to the client and reflect it in the download links. Every time anyone views a document, two new files are created, bloating disk consumption.

The third option, the good one, is to use sessions. Pleeeease! :)