Fixes #183. Fixes #175.
Split button only shown when a split is possible. It will pop up a new dialog which enumerates the splittable roles as a series of buttons. If there is only a single candidate, it will automatically be selected.

The requirement of having a type selected is removed from the span and arc dialogs, allowing misconfigured files to still modify the spans that do not match the configured types.