Closes #642 (possible targets defer/speedup)
Instead of using jQuery's selector, I switched here to
native JavaScript's `getElementsByClassName` to speed up
the search for the spans that need to be highlighted.
On my average-sized test document the routine speeds up by about 15x.

Also, the highlight is deferred to the drag instead of triggering
immediately on mousedown.