Updating the list of contributors
Anthony contributed his patch indirectly so I can't count his lines,
thus I add him to the end of the list. I hope no one really cares for
this SLOC silliness. Maybe we should just revert to alphabetic order but
I am scared if some user starts mailing the person at the top of the
list for support. If so, it better be someone that can at least be held
accountable for a large chunk of the code.