Implementing clientside for #216, removing UI
The discussion of significant issues relating to bulk upload
is left unresolved in #216. This results in serverside code
not being implemented. The presence of the UI element to
initiate bulk upload only serves to confuse the users at this
point, so I commented it out. However, I have done the groundwork
on the clientside so that file upload could be easily added one
day, when and if the issues are resolved and the server code does
get implemented.