only show relevant entity attributes on initial dialog display
Currently it seems that if you create a new text bound annotation then all possible entity attributes are shown when the dialog opens. If you change the entity type then this corrects the problem and restricts to just the relevant attributes. Moving the setting of showAllAttributes to false to before (rather than after) the calls to showAttributesFor fixes the problem (and seems to make more sense to me)